fixed protocol

+++ src/utils/ssl/recv.c
@@ -1,12 +1,11 @@
-#include "utils.h"
 #include "protocol.h"
 
 /*
  * Receive first packet with size of next packet
  */
-static int message_size(SSL *ssl) {
+static size_t message_size(SSL *ssl) {
     char buf[4];
-    size_t bytes = 0;
+    int bytes = 0;
     int size = -1;
 
     bytes = SSL_read(ssl, buf, sizeof(buf));
@@ -31,7 +30,7 @@ t_dtp *mx_recv(SSL *ssl) {
 
         buf[size] = '\0';
         if ((bytes = SSL_read(ssl, buf, sizeof(buf))) == size
-            && strlen(buf) == size)
+            && (int)strlen(buf) == size)
             dtp = mx_request_creation(buf);
         else
             mx_logger(MX_LOG_FILE, LOGWAR, "mx_recv\n");
